+++ benches/storage.rs
@@ -1,7 +1,7 @@
 use criterion::{black_box, criterion_group, criterion_main, BenchmarkId, Criterion};
 use hexx::{
     shapes::rombus,
-    storage::{HexStore, HexagonalMap, RombusMap},
+    storage::{HexModMap, HexStore, HexagonalMap, RombusMap},
     *,
 };
 use std::time::Duration;
@@ -15,7 +15,7 @@ pub fn hexagonal_map_benchmark(c: &mut Criterion) {
 
     let get_value = |h: Hex| h.length();
 
-    for dist in [10, 50, 100, 300] {
+    for dist in [10, 100, 300] {
         let std_hash_map: std::collections::HashMap<_, _> = Hex::ZERO
             .range(dist)
             .map(|hex| (hex, get_value(hex)))
@@ -25,6 +25,7 @@ pub fn hexagonal_map_benchmark(c: &mut Criterion) {
             .map(|hex| (hex, get_value(hex)))
             .collect();
         let hex_map = HexagonalMap::new(Hex::ZERO, dist, get_value);
+        let hexmod_map = HexModMap::new(Hex::ZERO, dist, get_value);
         group.bench_with_input(
             BenchmarkId::new("std::HashMap_get", dist),
             &dist,
@@ -58,6 +59,13 @@ pub fn hexagonal_map_benchmark(c: &mut Criterion) {
                 })
             },
         );
+        group.bench_with_input(BenchmarkId::new("HexModMap_get", dist), &dist, |b, dist| {
+            b.iter(|| {
+                for c in black_box(Hex::ZERO).range(*dist) {
+                    hexmod_map.get(c).unwrap();
+                }
+            })
+        });
         group.bench_with_input(
             BenchmarkId::new("std::HashMap_iter", dist),
             &dist,
@@ -73,6 +81,9 @@ pub fn hexagonal_map_benchmark(c: &mut Criterion) {
             &dist,
             |b, _| b.iter(|| hex_map.iter().collect::<Vec<_>>()),
         );
+        group.bench_with_input(BenchmarkId::new("HexModMap_iter", dist), &dist, |b, _| {
+            b.iter(|| hexmod_map.iter().collect::<Vec<_>>())
+        });
     }
     group.finish();
 }