+++ src/storage/mod.rs
@@ -7,9 +7,11 @@
 //!
 //! [this article]: https://www.redblobgames.com/grids/hexagons/#map-storage
 pub(crate) mod hexagonal;
+pub(crate) mod hexmod;
 pub(crate) mod rombus;
 
 pub use hexagonal::HexagonalMap;
+pub use hexmod::HexModMap;
 pub use rombus::RombusMap;
 
 macro_rules! storage_impl {
@@ -46,6 +48,7 @@ macro_rules! storage_impl {
 
 storage_impl!(HexagonalMap<T>);
 storage_impl!(RombusMap<T>);
+storage_impl!(HexModMap<T>);
 
 /// Trait grouping common features for hexagonal storage types.
 ///